requirements.txt
filetype==1.2.0
pandas==2.2.0
pygithub==2.2.0
PyLD
ruamel.yaml<0.18.0
ruamel.yaml.string
fuzzywuzzy
